requirements.txt
torch>=2.0.0
matplotlib
diffusers
transformers
accelerate
omegaconf
pillow
numpy
